refactored data fetching/parsing code
