+++ frontend/components/Viewing/PageJSON/Fetching/executeQueryFromTableJSON.js
@@ -1,0 +1,11 @@
+import getQueryResponse from '../../../../sparql/tools/getQueryResponse';
+import buildQuery from './buildQuery';
+
+export default function executeQueryFromTableJSON(uri, prefixes, table) {
+  return getQueryResponse(
+    prefixes + '\n' + buildQuery(uri, table),
+    {},
+    '',
+    false
+  );
+}
