+++ frontend/sparql/tools/getQueryResponse.js
@@ -6,8 +6,6 @@ const { publicRuntimeConfig } = getConfig();
 export default async function getQueryResponse(query, options, token, admin) {
   query = loadTemplate(query, options);
 
-  // console.log(query);
-
   const params = admin ? '/admin/sparql?query=' : '/sparql?query=';
   const graph = '&default-graph-uri=https://synbiohub.org/user/benjhatch7';
   const url = `${publicRuntimeConfig.backend}${params}${encodeURIComponent(
